+++ sdk/src/config.ts
@@ -11,6 +11,7 @@ import {
 	MainnetSpotMarkets,
 } from './constants/spotMarkets';
 import { OracleInfo } from './oracles/types';
+import { Program } from '@coral-xyz/anchor';
 
 type DriftConfig = {
 	ENV: DriftEnv;
@@ -110,6 +111,43 @@ export function getMarketsAndOraclesForSubscription(env: DriftEnv): {
 	return {
 		perpMarketIndexes: perpMarketIndexes,
 		spotMarketIndexes: spotMarketIndexes,
+		oracleInfos: Array.from(oracleInfos.values()),
+	};
+}
+
+export async function findAllMarketAndOracles(program: Program): Promise<{
+	perpMarketIndexes: number[];
+	spotMarketIndexes: number[];
+	oracleInfos: OracleInfo[];
+}> {
+	const perpMarketIndexes = [];
+	const spotMarketIndexes = [];
+	const oracleInfos = new Map<string, OracleInfo>();
+
+	const perpMarketProgramAccounts = await program.account.perpMarket.all();
+	const spotMarketProgramAccounts = await program.account.spotMarket.all();
+
+	for (const perpMarketProgramAccount of perpMarketProgramAccounts) {
+		const perpMarket = perpMarketProgramAccount.account;
+		perpMarketIndexes.push(perpMarket.marketIndex);
+		oracleInfos.set(perpMarket.amm.oracle.toString(), {
+			publicKey: perpMarket.amm.oracle,
+			source: perpMarket.amm.oracleSource,
+		});
+	}
+
+	for (const spotMarketProgramAccount of spotMarketProgramAccounts) {
+		const spotMarket = spotMarketProgramAccount.account;
+		spotMarketIndexes.push(spotMarket.marketIndex);
+		oracleInfos.set(spotMarket.oracle.toString(), {
+			publicKey: spotMarket.oracle,
+			source: spotMarket.oracleSource,
+		});
+	}
+
+	return {
+		perpMarketIndexes,
+		spotMarketIndexes,
 		oracleInfos: Array.from(oracleInfos.values()),
 	};
 }